+++ core/src/validations.ts
@@ -1,5 +1,5 @@
 import { RuntimeLink, RuntimeSubsystem, RuntimeSystem } from "./runtime.js";
-import { SystemMargin } from "./consts.js";
+import { SystemMargin } from "./helpers.js";
 import { System, specification } from "./specification.js";
 
 export class InvalidSystemError extends Error {
